MySQL refactor to long playerId and refactor back to int

+++ S05T01N01/src/main/java/cat/itacademy/s05/t01/n01/models/Player.java
@@ -15,7 +15,7 @@ public class Player {
 
 	@Id
 	@Column("playerId")
-	private long playerId;
+	private int playerId;
 	@Column("playerName")
 	private String playerName;
 	@Column("playerMaxPoints")
@@ -48,7 +48,7 @@ public class Player {
 		return Mono.fromRunnable(() -> this.playerName = playerName);
 	}
 
-	public Mono<Long> getPlayerId() {
+	public Mono<Integer> getPlayerId() {
 		return Mono.just(playerId);
 	}
